requirements.txt
openpyxl==3.0.5
bitstruct==8.11.0
boto3==1.16.10
jsonrpcclient==3.3.6
pyserial==3.4
